Moved methods to retrieve citations from NcbiUtils to NcbiPublicationSearchService.
