+++ panoramapublic/src/org/labkey/panoramapublic/proteomexchange/NcbiUtils.java
@@ -15,15 +15,10 @@
  */
 package org.labkey.panoramapublic.proteomexchange;
 
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.logging.log4j.Logger;
-import org.json.JSONException;
 import org.json.JSONObject;
 import org.labkey.api.collections.IntHashMap;
 import org.labkey.api.util.PageFlowUtil;
-import org.labkey.api.util.Pair;
-import org.jetbrains.annotations.Nullable;
 import org.labkey.api.util.logging.LogHelper;
 import org.labkey.api.view.AjaxCompletion;
 import org.w3c.dom.CharacterData;
@@ -38,7 +33,6 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
@@ -49,43 +43,16 @@ import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.apache.logging.log4j.Logger;
+
+/**
+ * Utilities for NCBI taxonomy lookups (autocomplete and scientific name resolution).
+ */
 public class NcbiUtils
 {
     private static final Pattern pattern = Pattern.compile("new Array\\(\"(.*)\"\\),");
     private static final String eutilsUrl = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esummary.fcgi?db=taxonomy";
     private static final String autoCompUrl = "https://blast.ncbi.nlm.nih.gov/portal/utils/autocomp.fcgi?dict=taxids_sg&q=";
-
-    private static final String pubmedCitationExporterUrl = "https://api.ncbi.nlm.nih.gov/api/ctxp/v1/pubmed/?format=citation&id=";
-    private static final String pmcCitationExporterUrl =    "https://api.ncbi.nlm.nih.gov/api/ctxp/v1/pmc/?format=citation&id=";
-    public static final String PUBMED_ID = "^[0-9]{1,8}$"; // https://libguides.library.arizona.edu/c.php?g=406096&p=2779570
-
-    public enum DB
-    {
-        PubMed("PubMed"),
-        PMC("PubMed Central");
-
-        private final String _label;
-
-        DB(String label)
-        {
-            _label = label;
-        }
-
-        public String getLabel()
-        {
-            return _label;
-        }
-
-        public static @Nullable DB fromString(@Nullable String database)
-        {
-            if (database == null) return null;
-            for (DB db : values())
-            {
-                if (db.name().equals(database)) return db;
-            }
-            return null;
-        }
-    }
 
     private static final Logger LOG = LogHelper.getLogger(NcbiUtils.class, "Messages about using the NCBI utilities");
 
@@ -209,96 +176,5 @@ public class NcbiUtils
             if(conn != null) conn.disconnect();
         }
         return sciNameMap;
-    }
-
-    /**
-     * Fetches the NLM-style citation for a publication using the NCBI Literature Citation Exporter.
-     * Supports both PubMed and PMC publication types.
-     * @param publicationId numeric publication ID (e.g. "28691345")
-     * @param database PubMed or PMC
-     * @return the NLM citation string, or null if the lookup fails or the ID is invalid
-     */
-    public static @Nullable String getCitation(String publicationId, DB database)
-    {
-        if (publicationId == null || !publicationId.matches(PUBMED_ID))
-        {
-            return null;
-        }
-
-        String baseUrl = database == DB.PMC ? pmcCitationExporterUrl : pubmedCitationExporterUrl;
-        String queryUrl = baseUrl + publicationId;
-
-        HttpURLConnection conn = null;
-        try
-        {
-            URL url = new URL(queryUrl);
-            conn = (HttpURLConnection) url.openConnection();
-            conn.setRequestMethod("GET");
-
-            int status = conn.getResponseCode();
-
-            if (status == HttpURLConnection.HTTP_OK)
-            {
-                String response;
-                try (InputStream in = conn.getInputStream())
-                {
-                    response = IOUtils.toString(in, StandardCharsets.UTF_8);
-                }
-                return parseCitation(response, publicationId, database);
-            }
-        }
-        catch (IOException e)
-        {
-            LOG.error("Error submitting a request to NCBI Literature Citation Exporter. URL: " + queryUrl, e);
-        }
-        finally
-        {
-            if(conn != null) conn.disconnect();
-        }
-        return null;
-    }
-
-    /**
-     * Does a citation lookup for the given PubMedId using the NCBI's Literature Citation Exporter (https://pmc.ncbi.nlm.nih.gov/api/ctxp).
-     * Example URL: https://api.ncbi.nlm.nih.gov/lit/ctxp/v1/pubmed/?format=citation&id=28691345
-     * @param pubmedId PubMed Id
-     * @return the PubMed link and the NLM-style citation if the lookup is successful
-     */
-    public static Pair<String, String> getPubMedLinkAndCitation(String pubmedId)
-    {
-        String citation = getCitation(pubmedId, DB.PubMed);
-        return citation != null ? new Pair<>(getPubmedLink(pubmedId), citation) : null;
-    }
-
-    public static String getPubmedLink(String pubmedId)
-    {
-        // Example: https://pubmed.ncbi.nlm.nih.gov/29331002
-        return "https://pubmed.ncbi.nlm.nih.gov/" + pubmedId;
-    }
-
-    public static String getPmcLink(String pubmedId)
-    {
-        return "https://www.ncbi.nlm.nih.gov/pmc/articles/" + pubmedId;
-    }
-
-    private static String parseCitation(String response, String publicationId, DB database)
-    {
-        try
-        {
-            var jsonObject = new JSONObject(response);
-            /* We are interested in the NLM style citation.
-            nlm: {
-                orig: "ENCODE Project Consortium. An integrated encyclopedia of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74. doi: 10.1038/nature11247. PMID: 22955616; PMCID: PMC3439153.",
-                format: "ENCODE Project Consortium. An integrated encyclopedia of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74. doi: 10.1038/nature11247. PMID: 22955616; PMCID: PMC3439153."
-            }
-            */
-            var nlmInfo = jsonObject.optJSONObject("nlm");
-            return null != nlmInfo ? nlmInfo.getString("orig") : null;
-        }
-        catch (JSONException e)
-        {
-            LOG.error("Error parsing response from NCBI Literature Citation Exporter for " + database.getLabel() + " ID " + publicationId, e);
-        }
-        return null;
     }
 }
